Backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework
django-cors-headers
mongoengine
python-dotenv
PyJWT
dnspython
sqlparse==0.5.3
tzdata==2025.2
